switch login api to AuthController
