+++ src/main/java/com/bjit/royalclub/royalclubfootball/controller/AuthController.java
@@ -1,18 +1,23 @@
 package com.bjit.royalclub.royalclubfootball.controller;
 
 import com.bjit.royalclub.royalclubfootball.model.ChangePasswordRequest;
+import com.bjit.royalclub.royalclubfootball.model.LoginRequest;
+import com.bjit.royalclub.royalclubfootball.model.LoginResponse;
 import com.bjit.royalclub.royalclubfootball.service.AuthService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.LOGIN_OK;
 import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.UPDATE_OK;
 import static com.bjit.royalclub.royalclubfootball.util.ResponseBuilder.buildResponse;
+import static com.bjit.royalclub.royalclubfootball.util.ResponseBuilder.buildSuccessResponse;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,6 +25,12 @@ import static com.bjit.royalclub.royalclubfootball.util.ResponseBuilder.buildRes
 public class AuthController {
 
     private final AuthService authService;
+
+    @PostMapping("/login")
+    public ResponseEntity<Object> login(@Valid @RequestBody LoginRequest loginRequest) {
+        LoginResponse loginResponse = authService.login(loginRequest);
+        return buildSuccessResponse(HttpStatus.OK, LOGIN_OK, loginResponse);
+    }
 
     @PutMapping("/change-password")
     public ResponseEntity<Object> changePassword(@Valid @RequestBody ChangePasswordRequest changePasswordRequest) {
