+++ src/main/java/com/bjit/royalclub/royalclubfootball/controller/PlayerController.java
@@ -1,7 +1,5 @@
 package com.bjit.royalclub.royalclubfootball.controller;
 
-import com.bjit.royalclub.royalclubfootball.model.LoginRequest;
-import com.bjit.royalclub.royalclubfootball.model.LoginResponse;
 import com.bjit.royalclub.royalclubfootball.model.PlayerRegistrationRequest;
 import com.bjit.royalclub.royalclubfootball.model.PlayerResponse;
 import com.bjit.royalclub.royalclubfootball.model.PlayerUpdateRequest;
@@ -24,7 +22,6 @@ import java.util.List;
 
 import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.CREATE_OK;
 import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.FETCH_OK;
-import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.LOGIN_OK;
 import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.STATUS_UPDATE_OK;
 import static com.bjit.royalclub.royalclubfootball.constant.RestResponseMessage.UPDATE_OK;
 import static com.bjit.royalclub.royalclubfootball.util.ResponseBuilder.buildSuccessResponse;
@@ -69,9 +66,4 @@ public class PlayerController {
         return buildSuccessResponse(HttpStatus.OK, UPDATE_OK, playerResponse);
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<Object> login(@Valid @RequestBody LoginRequest loginRequest) {
-        LoginResponse loginResponse = playerService.login(loginRequest);
-        return buildSuccessResponse(HttpStatus.OK, LOGIN_OK, loginResponse);
-    }
 }
