+++ src/main/java/com/bjit/royalclub/royalclubfootball/service/PlayerService.java
@@ -1,8 +1,6 @@
 package com.bjit.royalclub.royalclubfootball.service;
 
 import com.bjit.royalclub.royalclubfootball.entity.Player;
-import com.bjit.royalclub.royalclubfootball.model.LoginRequest;
-import com.bjit.royalclub.royalclubfootball.model.LoginResponse;
 import com.bjit.royalclub.royalclubfootball.model.PlayerRegistrationRequest;
 import com.bjit.royalclub.royalclubfootball.model.PlayerResponse;
 import com.bjit.royalclub.royalclubfootball.model.PlayerUpdateRequest;
@@ -25,8 +23,5 @@ public interface PlayerService {
     PlayerResponse updatePlayer(Long id, PlayerUpdateRequest updateRequest);
 
     Player findByEmail(String userName);
-
-    LoginResponse login(LoginRequest loginRequest);
-
 
 }
